cast std::size_type to unsigned long and use %ul so types match on all platforms -DB

+++ util/mp4chaps.cpp
@@ -306,7 +306,7 @@ ChapterUtility::actionEvery( JobContext& job )
     {
         MP4Chapter_t chap;
         chap.duration = refTrackDuration.duration > chapterDuration.duration ? chapterDuration.duration : refTrackDuration.duration;
-        sprintf(chap.title, "Chapter %u", chapters.size()+1);
+        sprintf(chap.title, "Chapter %lu", (unsigned long)chapters.size()+1);
 
         chapters.push_back( chap );
         refTrackDuration -= chapterDuration;
